Handle missing support for Blivet-GUI in the Storage module (#1685645)
The Blivet partitioning module should be able to exist without support
for Blivet-GUI. In this case, the method SendRequest will raise the
UnsupportedPartitioningError exception.

In the future, we will raise this exception once the Blivet
partitioning is requested.

Resolves: rhbz#1685645
